+++ compiler/src/CLI_errors.ml
@@ -56,7 +56,7 @@ let check_options () =
 
   if !latexfile <> ""
   then warning Deprecated Location.i_dummy
-         "the [-latex] option has been deprecated since March 2023; use [jazz2tex] instead";
+         "the [-latex] option has been deprecated since March 2023; use [jasmin2tex] instead";
 
   if Option.is_some !ct_list || !infer
   then warning Deprecated Location.i_dummy